backend/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cloudinary==1.42.2
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
Django==4.2.17
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
multidict==6.1.0
phonenumbers==9.0.0
pillow==11.1.0
propcache==0.3.0
psycopg==3.2.5
psycopg2-binary>=2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
twilio==9.4.6
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zope.interface==7.2
psycopg2-binary>=2.9